transcoder: Move GetClaimManager into transcoder.go.
Also document the logical layout of the code within LivepeerNode.
