+++ core/transcoder.go
@@ -325,6 +325,28 @@ func (n *LivepeerNode) transcodeSegmentLoop(job *ethTypes.Job, segChan SegmentCh
 	return nil
 }
 
+func (n *LivepeerNode) GetClaimManager(job *ethTypes.Job) (eth.ClaimManager, error) {
+	n.claimMutex.Lock()
+	defer n.claimMutex.Unlock()
+	if job == nil {
+		glog.Error("Nil job")
+		return nil, fmt.Errorf("Nil job")
+	}
+	jobId := job.JobId.Int64()
+	// XXX we should clear entries after some period of inactivity
+	if cm, ok := n.ClaimManagers[jobId]; ok {
+		return cm, nil
+	}
+	// no claimmanager exists yet; check if we're assigned the job
+	if n.Eth == nil {
+		return nil, nil
+	}
+	glog.Infof("Creating new claim manager for job %v", jobId)
+	cm := eth.NewBasicClaimManager(job, n.Eth, n.Ipfs, n.Database)
+	n.ClaimManagers[jobId] = cm
+	return cm, nil
+}
+
 func randName() string {
 	rand.Seed(time.Now().UnixNano())
 	x := make([]byte, 10, 10)
